+++ mne/icons/resources.py
