--- mne/images/mne.qrc
@@ -1,14 +1,0 @@
-<!DOCTYPE RCC><RCC version="1.0">
-<qresource>
-    <file alias="visibility_on.svg">mne/images/visibility_on-black-18dp.svg</file>
-    <file alias="visibility_off.svg">mne/images/visibility_off-black-18dp.svg</file>
-    <file alias="help.svg">mne/images/help-black-18dp.svg</file>
-    <file alias="play.svg">mne/images/play-black-18dp.svg</file>
-    <file alias="pause.svg">mne/images/pause-black-18dp.svg</file>
-    <file alias="scale.svg">mne/images/scale-black-18dp.svg</file>
-    <file alias="restore.svg">mne/images/restore-black-18dp.svg</file>
-    <file alias="clear.svg">mne/images/clear-black-18dp.svg</file>
-    <file alias="screenshot.svg">mne/images/screenshot-black-18dp.svg</file>
-    <file alias="movie.svg">mne/images/movie-black-18dp.svg</file>
-</qresource>
-</RCC>
